Capstone Part 2: Question 1
Question 1
Make changes to settings.py to enable the use of Django Configurations. Create a class calledDev to hold the settings. Note that you don’t need to change manage.py as it has already been configured to work with both class-based and standard settings.
Then, add a LOGGING config. Add a console handle that prints to STDOUT, and make sure the root uses this handler with a log level of DEBUG.