Capstone Part 2: Question 2
Question 2
In this question you’ll make a number of changes to support Django Rest Framework. The first is to configure the settings. Here are the REST_FRAMEWORK settings that we want:
Settings
DEFAULT_PERMISSION_CLASSES should be rest_framework.permissions.IsAuthenticated.
A user should be able to authenticate with BasicAuthentication, SessionAuthentication, TokenAuthentication and JWTAuthentication.
rest_framework.pagination.PageNumberPagination should be used as the default pagination class.