Capstone Part 2: Question 3
Question 3
We want to have custom permissions on MovieNightViewSet and MovieNightInvitationViewSet, to allow changes only by the “owner” of their respective models. Each of these models stores the “owner” on different fields. For MovieNight, it’s the creator, and for MovieNightInvitation, the invitee. In movies/api/permisions.py the UserFieldPermission class has been created. It checks for permission by reading the owner from the user_field attribute of the model instance.
IsCreatorPermission and IsInviteePermission need to be completed to implement the correct user_field attribute for their respective models.
Once they’re complete, apply each to the MovieNightViewSet and MovieNightInvitationViewSet, using the permission_classes. rest_framework.permissions.IsAuthenticated should also be applied as a permission to each of these classes; remember that permission classes can be combined with the bitwise AND (&) operator.