Capstone Part 2: Question 4
Question 4
Sometimes it could take a while to send out notification emails, so we want to use Celery for this. The Celery app is already set up for the project, some settings just need to be configured.
First changes need to be made to INSTALLED_APPS to store the results in the Django DB. Then, CELERY_RESULT_BACKEND and CELERY_BROKER_URL need to be configured appropriately. You will have to migrate after this.
MIGRATE
Task functions have been set written in movies/tasks.py but they have not been configured to be run as Celery tasks. Import the right decorator function and use it on send_invitation, send_attendance_change and notify_of_starting_soon.
After you have done this, USE_CELERY in movies/signals.py can be set to True.